Added configs for logging, architecture corrections
Added logging config, provided new utils, added type hints to entities, started working with handlers
