+++ app/domain/entities/user.py
@@ -9,11 +9,12 @@ class User:
     def update(self, field, value):
         self.data[field] = value
 
-    def to_dict(self):
+    def to_dict(self) -> dict:
         return self.data
     
     @classmethod
-    def from_dict(cls, id, data_dict):
+    def from_dict(cls, id, data_dict) -> 'User':
+
         user = cls(id)
         user.data = data_dict
         return user
