--- app/use_cases/start.py
@@ -1,4 +1,0 @@
-from entities.user import User
-
-
-
