+++ presentation/messages.py
@@ -1,0 +1,3 @@
+START_MSG = '''Hello {username}\! It's a bot for creating music quizes\. \
+                            Do not forget to challenge your friends\!\n\n'''
+
